+++ src/activation_functions.rs
@@ -9,7 +9,7 @@ pub enum ActivationFunction {
 }
 
 impl ActivationFunction {
-    pub fn apply(&self, weighted_inputs: &Array2<f64>) -> Array2<f64> {
+    pub fn apply(&self, weighted_inputs: &Array2<f32>) -> Array2<f32> {
         match *self {
             ActivationFunction::Id => weighted_inputs.clone(),
             ActivationFunction::Sigmoid => {
@@ -17,42 +17,42 @@ impl ActivationFunction {
                     .iter()
                     .map(|x| 1.0 / (1.0 + (-*x).exp()))
                     .collect();
-                Array2::<f64>::from_shape_vec(weighted_inputs.dim(), elems).unwrap()
+                Array2::<f32>::from_shape_vec(weighted_inputs.dim(), elems).unwrap()
             }
             ActivationFunction::Relu => {
                 let elems: Vec<_> = weighted_inputs.iter().map(|x| x.max(0.0)).collect();
-                Array2::<f64>::from_shape_vec(weighted_inputs.dim(), elems).unwrap()
+                Array2::<f32>::from_shape_vec(weighted_inputs.dim(), elems).unwrap()
             }
             ActivationFunction::SoftMax => {
-                let col_exp_sums: Vec<f64> = weighted_inputs
+                let col_exp_sums: Vec<f32> = weighted_inputs
                     .columns()
                     .into_iter()
                     .map(|col| col.iter().map(|x| x.exp()).sum())
                     .collect();
-                Array2::<f64>::from_shape_fn(weighted_inputs.dim(), |(i, j)| {
-                    weighted_inputs[(i, j)].exp() * (1_f64 / col_exp_sums[j])
+                Array2::<f32>::from_shape_fn(weighted_inputs.dim(), |(i, j)| {
+                    weighted_inputs[(i, j)].exp() * (1_f32 / col_exp_sums[j])
                 })
             }
         }
     }
 
-    pub fn derivative(&self, weighted_inputs: &Array2<f64>) -> Array2<f64> {
+    pub fn derivative(&self, weighted_inputs: &Array2<f32>) -> Array2<f32> {
         match *self {
-            ActivationFunction::Id => Array2::<f64>::from_elem(weighted_inputs.dim(), 1.0),
+            ActivationFunction::Id => Array2::<f32>::from_elem(weighted_inputs.dim(), 1.0),
             ActivationFunction::Sigmoid => {
                 let elems: Vec<_> = weighted_inputs.iter().map(|x| *x * (1.0 - *x)).collect();
-                Array2::<f64>::from_shape_vec(weighted_inputs.dim(), elems).unwrap()
+                Array2::<f32>::from_shape_vec(weighted_inputs.dim(), elems).unwrap()
             }
             ActivationFunction::Relu => {
                 let elems: Vec<_> = weighted_inputs
                     .iter()
                     .map(|x| if *x > 0.0 { 1.0 } else { 0.0 })
                     .collect();
-                Array2::<f64>::from_shape_vec(weighted_inputs.dim(), elems).unwrap()
+                Array2::<f32>::from_shape_vec(weighted_inputs.dim(), elems).unwrap()
             }
             ActivationFunction::SoftMax => {
                 let softmax = self.apply(weighted_inputs);
-                let ones = Array2::<f64>::from_elem(weighted_inputs.dim(), 1_f64);
+                let ones = Array2::<f32>::from_elem(weighted_inputs.dim(), 1_f32);
                 &softmax * (ones - &softmax)
             }
         }
@@ -82,13 +82,13 @@ impl ActivationFunction {
 }
 
 fn xavier_glorot_normal(
-    denominator: f64,
+    denominator: f32,
     prev_layer_neuron_count: usize,
     neuron_count: usize,
 ) -> distribution::Distribution {
     distribution::Distribution::Normal {
         mean: 0.0,
-        standard_deviation: (denominator / prev_layer_neuron_count as f64).sqrt(),
+        standard_deviation: (denominator / prev_layer_neuron_count as f32).sqrt(),
     }
 }
 
@@ -97,7 +97,7 @@ fn xavier_glorot_uniform(
     neuron_count: usize,
 ) -> distribution::Distribution {
     let neurons_in_and_out = prev_layer_neuron_count + neuron_count;
-    let range_arond_zero = (6_f64 / neurons_in_and_out as f64).sqrt();
+    let range_arond_zero = (6_f32 / neurons_in_and_out as f32).sqrt();
     distribution::Distribution::Uniform {
         lower_bound: -range_arond_zero,
         upper_bound: range_arond_zero,
