+++ src/distribution.rs
@@ -1,4 +1,4 @@
 pub enum Distribution {
-    Normal { mean: f64, standard_deviation: f64 },
-    Uniform { lower_bound: f64, upper_bound: f64 },
+    Normal { mean: f32, standard_deviation: f32 },
+    Uniform { lower_bound: f32, upper_bound: f32 },
 }
